LEL-012 refactor: update journal model and dao to support long ids and filtering

+++ feature/menu/diary/src/main/java/io/github/faening/lello/feature/menu/diary/DiaryViewModel.kt
@@ -51,7 +51,7 @@ class DiaryViewModel @Inject constructor(
         }
     }
 
-    fun toggleDiaryStatus(diaryId: Int, active: Boolean) {
+    fun toggleDiaryStatus(diaryId: Long, active: Boolean) {
 //        viewModelScope.launch {
 //            try {
 //                val success = diaryRepository.toggleActiveStatus(diaryId, active)
